+++ classes/asserters/phpFloat.php
@@ -43,12 +43,17 @@ class phpFloat extends asserters\integer
 
 			switch (true)
 			{
+				case $absCurrentValue == 0 && $absValue == 0:
+					return $this->isEqualTo($value);
+
 				case $offsetIsNaN === true:
 				case $offset / ($absCurrentValue + $absValue) >= $epsilon:
 				case $absCurrentValue * $absValue == 0 && $offset >= pow($epsilon, 2):
 					$this->fail(($failMessage ?: $this->_('%s is not nearly equal to %s with epsilon %s', $this, $this->getTypeOf($value), $epsilon)) . PHP_EOL . $this->diff($value));
 			}
 		}
+
+		$this->pass();
 
 		return $this;
 	}